+++ lib/src/controls/navigation/navigation_view/style.dart
@@ -108,6 +108,8 @@ class NavigationPaneThemeData with Diagnosticable {
   final TextStyle? itemHeaderTextStyle;
   final ButtonState<TextStyle?>? selectedTextStyle;
   final ButtonState<TextStyle?>? unselectedTextStyle;
+  final ButtonState<TextStyle?>? selectedTopTextStyle;
+  final ButtonState<TextStyle?>? unselectedTopTextStyle;
   final ButtonState<Color?>? selectedIconColor;
   final ButtonState<Color?>? unselectedIconColor;
 
@@ -124,6 +126,8 @@ class NavigationPaneThemeData with Diagnosticable {
     this.itemHeaderTextStyle,
     this.selectedTextStyle,
     this.unselectedTextStyle,
+    this.selectedTopTextStyle,
+    this.unselectedTopTextStyle,
     this.animationDuration,
     this.animationCurve,
     this.selectedIconColor,
@@ -163,6 +167,24 @@ class NavigationPaneThemeData with Diagnosticable {
                   : resources.textFillColorPrimary,
         );
       }),
+      selectedTopTextStyle: ButtonState.resolveWith((states) {
+        return typography.body?.copyWith(
+          color: states.isPressing
+              ? resources.textFillColorTertiary
+              : states.isHovering
+                  ? resources.textFillColorSecondary
+                  : resources.textFillColorPrimary,
+        );
+      }),
+      unselectedTopTextStyle: ButtonState.resolveWith((states) {
+        return typography.body?.copyWith(
+          color: states.isPressing
+              ? resources.textFillColorSecondary
+              : states.isDisabled
+                  ? resources.textFillColorDisabled
+                  : resources.textFillColorPrimary,
+        );
+      }),
       labelPadding: const EdgeInsetsDirectional.only(end: 10.0),
       iconPadding: const EdgeInsets.symmetric(horizontal: 10.0),
       headerPadding: const EdgeInsetsDirectional.only(top: 10.0),
@@ -188,6 +210,10 @@ class NavigationPaneThemeData with Diagnosticable {
           a?.selectedTextStyle, b?.selectedTextStyle, t, TextStyle.lerp),
       unselectedTextStyle: ButtonState.lerp(
           a?.unselectedTextStyle, b?.unselectedTextStyle, t, TextStyle.lerp),
+      selectedTopTextStyle: ButtonState.lerp(
+          a?.selectedTextStyle, b?.selectedTextStyle, t, TextStyle.lerp),
+      unselectedTopTextStyle: ButtonState.lerp(
+          a?.unselectedTextStyle, b?.unselectedTextStyle, t, TextStyle.lerp),
       highlightColor: Color.lerp(a?.highlightColor, b?.highlightColor, t),
       animationCurve: t < 0.5 ? a?.animationCurve : b?.animationCurve,
       animationDuration: lerpDuration(a?.animationDuration ?? Duration.zero,
@@ -209,6 +235,9 @@ class NavigationPaneThemeData with Diagnosticable {
       itemHeaderTextStyle: style?.itemHeaderTextStyle ?? itemHeaderTextStyle,
       selectedTextStyle: style?.selectedTextStyle ?? selectedTextStyle,
       unselectedTextStyle: style?.unselectedTextStyle ?? unselectedTextStyle,
+      selectedTopTextStyle: style?.selectedTopTextStyle ?? selectedTopTextStyle,
+      unselectedTopTextStyle:
+          style?.unselectedTopTextStyle ?? unselectedTopTextStyle,
       highlightColor: style?.highlightColor ?? highlightColor,
       animationCurve: style?.animationCurve ?? animationCurve,
       animationDuration: style?.animationDuration ?? animationDuration,
@@ -234,6 +263,8 @@ class NavigationPaneThemeData with Diagnosticable {
       ..add(DiagnosticsProperty<Curve>('animationCurve', animationCurve))
       ..add(DiagnosticsProperty('selectedTextStyle', selectedTextStyle))
       ..add(DiagnosticsProperty('unselectedTextStyle', unselectedTextStyle))
+      ..add(DiagnosticsProperty('selectedTopTextStyle', selectedTextStyle))
+      ..add(DiagnosticsProperty('unselectedTopTextStyle', unselectedTextStyle))
       ..add(DiagnosticsProperty('selectedIconColor', selectedIconColor))
       ..add(DiagnosticsProperty('unselectedIconColor', unselectedIconColor));
   }